+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Plutus/Context.hs
@@ -1,17 +1,21 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE PolyKinds #-}
+{-# LANGUAGE QuantifiedConstraints #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilyDependencies #-}
+{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 -- Recursive definition constraints of `EraPlutusContext` and `EraPlutusTxInfo` lead to a wrongful
@@ -28,6 +32,7 @@
 module Cardano.Ledger.Alonzo.Plutus.Context (
   CollectError (..),
   LedgerTxInfo (..),
+  LevelTxInfo (..),
   EraPlutusTxInfo (..),
   PlutusTxInfoResult (..),
   mkPlutusTxInfoFromResult,
@@ -63,7 +68,7 @@ import Cardano.Ledger.Alonzo.Scripts (
   PlutusPurpose,
   PlutusScript (..),
  )
-import Cardano.Ledger.BaseTypes (ProtVer (..), Version, kindObjectValue)
+import Cardano.Ledger.BaseTypes (ProtVer (..), TxIx, Version, kindObjectValue)
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Core
@@ -92,6 +97,7 @@ import Data.Aeson (ToJSON (..), (.=), pattern String)
 import Data.Kind (Type)
 import Data.List.NonEmpty (NonEmpty, nonEmpty)
 import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
 import Data.Text (Text)
 import GHC.Generics
 import GHC.Stack
@@ -99,6 +105,25 @@ import qualified PlutusLedgerApi.V1 as PV1
 import qualified PlutusLedgerApi.V2 as PV2
 import qualified PlutusLedgerApi.V3 as PV3
 import qualified PlutusLedgerApi.V4 as PV4
+
+data LevelTxInfo (level :: TxLevel) era where
+  TopTxInfo ::
+    -- | This is a tricky field that is only used starting with Dijkstra era and only by top level
+    -- transactions. It is always safe to leave it as `mempty` upon construction
+    Map.Map TxId (TxInfoResult era) ->
+    LevelTxInfo TopTx era
+  SubTxInfo ::
+    !TxIx -> LevelTxInfo SubTx era
+
+instance
+  ( EraTxLevel era
+  , STxLevel TopTx era ~ STxBothLevels TopTx era
+  , STxLevel SubTx era ~ STxBothLevels SubTx era
+  ) =>
+  HasEraTxLevel LevelTxInfo era
+  where
+  toSTxLevel TopTxInfo {} = STopTx
+  toSTxLevel SubTxInfo {} = SSubTx
 
 -- | All information that is necessary from the ledger to construct Plutus' TxInfo.
 data LedgerTxInfo era where
@@ -108,9 +133,7 @@ data LedgerTxInfo era where
     , ltiSystemStart :: !SystemStart
     , ltiUTxO :: !(UTxO era)
     , ltiTx :: !(Tx level era)
-    , ltiMemoizedSubTransactions :: Map TxId (TxInfoResult era)
-    -- ^ This is a tricky field that is only used starting with Dijkstra era and only by top level
-    -- transactions. It is always safe to leave it as `mempty` upon construction, even for Dijkstra
+    , ltiLevelInfo :: LevelTxInfo level era
     } ->
     LedgerTxInfo era
 
@@ -168,7 +191,7 @@ class
     proxy l ->
     TxOutSource ->
     TxOut era ->
-    Either (ContextError era) (PlutusTxOut l)
+    Either (ContextError era) (PlutusTxOut era l)
 
 -- | This is the helper type that captures translation of `Tx` to `PlutusTxInfo`.
 --
@@ -295,13 +318,14 @@ type family PlutusRedeemerPointer (l :: Language) where
   PlutusRedeemerPointer 'PlutusV3 = (PlutusScriptPurpose 'PlutusV3, PV3.Redeemer)
   PlutusRedeemerPointer 'PlutusV4 = (PlutusScriptPurpose 'PlutusV4, PV4.Redeemer)
 
-type family PlutusTxOut (l :: Language) where
+type family PlutusTxOut era (l :: Language) where
   -- This special case is here because Alonzo does not have a ContextError
   -- for the case where it encounters a Byron address in a TxIn
-  PlutusTxOut 'PlutusV1 = Maybe PV1.TxOut
-  PlutusTxOut 'PlutusV2 = PV2.TxOut
-  PlutusTxOut 'PlutusV3 = PV3.TxOut
-  PlutusTxOut 'PlutusV4 = PV4.TxOut
+  PlutusTxOut AlonzoEra PlutusV1 = Maybe PV1.TxOut
+  PlutusTxOut _ PlutusV1 = PV1.TxOut
+  PlutusTxOut _ PlutusV2 = PV2.TxOut
+  PlutusTxOut _ PlutusV3 = PV3.TxOut
+  PlutusTxOut _ PlutusV4 = PV4.TxOut
 
 -- | This is just like `mkPlutusScript`, except it is guaranteed to be total through the enforcement
 -- of support by the type system and `EraPlutusTxInfo` type class instances for supported plutus