+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxInfo.hs
@@ -39,6 +39,7 @@ module Cardano.Ledger.Conway.TxInfo (
   transTxCertV1V2,
   transPlutusPurposeV1V2,
   transPlutusPurposeV3,
+  transSlotToPOSIXTime,
   guardConwayFeaturesForPlutusV1V2,
   transTxInInfoV3,
   scriptPurposeToScriptInfo,
@@ -72,6 +73,7 @@ import qualified Cardano.Ledger.Babbage.TxInfo as Babbage
 import Cardano.Ledger.BaseTypes (
   Inject (..),
   ProtVer (..),
+  SlotNo,
   StrictMaybe (..),
   getVersion32,
   isSJust,
@@ -454,7 +456,7 @@ instance EraPlutusTxInfo 'PlutusV1 ConwayEra where
 
   toPlutusRedeemerPointer = Alonzo.transRedeemerPointerV1
 
-  toPlutusTxOut _ src txOut = Just <$> transTxOutV1 src txOut
+  toPlutusTxOut _ = transTxOutV1
 
 instance EraPlutusTxInfo 'PlutusV2 ConwayEra where
   toPlutusTxCert _ _ = transTxCertV1V2
@@ -822,32 +824,40 @@ class
 instance ConwayEraPlutusTxInfo 'PlutusV3 ConwayEra where
   toPlutusChangedParameters _ x = PV3.ChangedParameters (PV3.dataToBuiltinData (toPlutusData x))
 
+transSlotToPOSIXTime ::
+  forall era proxy.
+  Inject (AlonzoContextError era) (ContextError era) =>
+  proxy era ->
+  EpochInfo (Either Text) ->
+  SystemStart ->
+  SlotNo ->
+  Either (ContextError era) PV3.POSIXTime
+transSlotToPOSIXTime _ epochInfo systemStart =
+  left (inject . TimeTranslationPastHorizon @era)
+    . slotToPOSIXTime epochInfo systemStart
+
 -- | Translate a validity interval to POSIX time
 transValidityInterval ::
-  forall proxy era a.
-  Inject (AlonzoContextError era) a =>
+  forall proxy era.
+  Inject (AlonzoContextError era) (ContextError era) =>
   proxy era ->
   EpochInfo (Either Text) ->
   SystemStart ->
   ValidityInterval ->
-  Either a PV1.POSIXTimeRange
-transValidityInterval _ epochInfo systemStart = \case
+  Either (ContextError era) PV1.POSIXTimeRange
+transValidityInterval era epochInfo systemStart = \case
   ValidityInterval SNothing SNothing -> pure PV1.always
-  ValidityInterval (SJust i) SNothing -> PV1.from <$> transSlotToPOSIXTime i
+  ValidityInterval (SJust i) SNothing -> PV1.from <$> transSlotToPOSIXTime era epochInfo systemStart i
   ValidityInterval SNothing (SJust i) -> do
-    t <- transSlotToPOSIXTime i
+    t <- transSlotToPOSIXTime era epochInfo systemStart i
     pure $ PV1.Interval (PV1.LowerBound PV1.NegInf True) (PV1.strictUpperBound t)
   ValidityInterval (SJust i) (SJust j) -> do
-    t1 <- transSlotToPOSIXTime i
-    t2 <- transSlotToPOSIXTime j
+    t1 <- transSlotToPOSIXTime era epochInfo systemStart i
+    t2 <- transSlotToPOSIXTime era epochInfo systemStart j
     pure $
       PV1.Interval
         (PV1.lowerBound t1)
         (PV1.strictUpperBound t2)
-  where
-    transSlotToPOSIXTime =
-      left (inject . TimeTranslationPastHorizon @era)
-        . slotToPOSIXTime epochInfo systemStart
 
 checkReferenceInputsNotDisjointFromInputs ::
   forall l era.