+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/TxInfo.hs
@@ -1,9 +1,10 @@
-{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GADTs #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
@@ -36,16 +37,17 @@ import Cardano.Ledger.Alonzo.Plutus.Context (
   EraPlutusContext (..),
   EraPlutusTxInfo (..),
   LedgerTxInfo (..),
+  LevelTxInfo (..),
   PlutusPurposeScriptHashArg,
   PlutusRedeemerPointer,
   PlutusScriptPurpose,
+  PlutusTxInfo,
   PlutusTxInfoResult (..),
   SupportedLanguage (..),
   SupportedPlutusRunnable (..),
  )
-import Cardano.Ledger.Alonzo.Plutus.TxInfo (transPolicyID, transValue)
 import qualified Cardano.Ledger.Alonzo.Plutus.TxInfo as Alonzo
-import Cardano.Ledger.Alonzo.Scripts (toAsItem)
+import Cardano.Ledger.Alonzo.Scripts (AsPurpose, toAsItem, toAsPurpose)
 import Cardano.Ledger.Alonzo.UTxO (AlonzoEraUTxO (..))
 import Cardano.Ledger.Babbage.TxInfo (BabbageContextError (..), transRedeemer, transReferenceScript)
 import qualified Cardano.Ledger.Babbage.TxInfo as Babbage
@@ -55,7 +57,7 @@ import Cardano.Ledger.BaseTypes (
   Inject (..),
   ProtVer (..),
   StrictMaybe (..),
-  TxIx (TxIx),
+  TxIx (..),
   kindObjectValue,
   strictMaybe,
   strictMaybeToMaybe,
@@ -74,6 +76,7 @@ import Cardano.Ledger.Conway.TxInfo (
   transHotCommitteeCred,
   transMap,
   transProposal,
+  transSlotToPOSIXTime,
   transTxInInfoV1,
   transTxInInfoV3,
   transVoter,
@@ -103,6 +106,7 @@ import Cardano.Ledger.Plutus (
   decodePlutusRunnable,
   getPlutusData,
   plutusLanguage,
+  transAccountAddress,
   transCoinToLovelace,
   transCoinToValue,
   transCred,
@@ -115,8 +119,10 @@ import Cardano.Ledger.Plutus (
  )
 import Cardano.Ledger.Plutus.Data (Data)
 import Cardano.Ledger.Plutus.ToPlutusData (ToPlutusData (..))
-import Cardano.Ledger.State (StakePoolParams (..), UTxO)
+import Cardano.Ledger.Slot (EpochInfo)
+import Cardano.Ledger.State (EraUTxO (..), StakePoolParams (..), UTxO)
 import Cardano.Ledger.TxIn (TxId (TxId), TxIn (..))
+import Cardano.Slotting.Time (SystemStart)
 import Control.DeepSeq (NFData)
 import Control.Monad (unless, zipWithM)
 import Data.Aeson (KeyValue (..), ToJSON (..))
@@ -132,12 +138,15 @@ import Data.Proxy (Proxy (..))
 import qualified Data.Set as Set
 import Data.Set.NonEmpty (NonEmptySet)
 import qualified Data.Set.NonEmpty as NES
+import Data.Text (Text)
 import GHC.Generics (Generic)
 import Lens.Micro ((^.))
 import qualified PlutusLedgerApi.V1 as PV1
 import qualified PlutusLedgerApi.V2 as PV2
 import qualified PlutusLedgerApi.V3 as PV3
 import qualified PlutusLedgerApi.V4 as PV4
+import qualified PlutusTx.AssocMap ()
+import qualified PlutusTx.AssocMap as AssocMap
 
 data DijkstraContextError era
   = ConwayContextError (ConwayContextError era)
@@ -359,7 +368,7 @@ instance EraPlutusTxInfo 'PlutusV1 DijkstraEra where
 
   toPlutusRedeemerPointer = Alonzo.transRedeemerPointerV1
 
-  toPlutusTxOut _ src txOut = Just <$> Conway.transTxOutV1 src txOut
+  toPlutusTxOut _ = Conway.transTxOutV1
 
 transTxCertV1V2 ::
   ( ConwayEraTxCert era
@@ -503,7 +512,7 @@ guardDijkstraFeaturesForPlutusV1toV3 tx = do
   let txBody = tx ^. bodyTxL
       directDeposits = txBody ^. directDepositsTxBodyL
       accountBalanceIntervals = txBody ^. accountBalanceIntervalsTxBodyL
-      requiredTopLevelGuards = txBody ^. requiredTopLevelGuardsL
+      requiredTopLevelGuards = txBody ^. requiredTopLevelGuardsTxBodyL
       scriptHashes = [sh | ScriptHashObj sh <- toList (txBody ^. guardsTxBodyL)]
   unless (null $ unDirectDeposits directDeposits) $
     Left $
@@ -610,11 +619,11 @@ instance EraPlutusTxInfo 'PlutusV4 DijkstraEra where
   toPlutusTxInfo proxy LedgerTxInfo {..} =
     PlutusTxInfoResult $ do
       let
+        era = Proxy @DijkstraEra
         txBody = ltiTx ^. bodyTxL
         txInputs = txBody ^. inputsTxBodyL
         refInputs = txBody ^. referenceInputsTxBodyL
-      timeRange <-
-        Conway.transValidityInterval ltiTx ltiEpochInfo ltiSystemStart (txBody ^. vldtTxBodyL)
+      timeRange <- transValidityInterval era ltiEpochInfo ltiSystemStart (txBody ^. vldtTxBodyL)
       inputsInfo <- mapM (transTxInInfoV4 ltiUTxO) (Set.toList txInputs)
       refInputsInfo <- mapM (transTxInInfoV4 ltiUTxO) (Set.toList refInputs)
       Conway.checkReferenceInputsNotDisjointFromInputs txBody
@@ -649,13 +658,11 @@ instance EraPlutusTxInfo 'PlutusV4 DijkstraEra where
             { PV4.txInfoInputs = inputsInfo
             , PV4.txInfoOutputs = outputs
             , PV4.txInfoReferenceInputs = refInputsInfo
-            , PV4.txInfoFee =
-                withBothTxLevels txBody (\topTxBody -> transCoinToLovelace (topTxBody ^. feeTxBodyL)) (const 0)
             , PV4.txInfoMint = Conway.transMintValue (txBody ^. mintTxBodyL)
             , PV4.txInfoTxCerts = txCerts
             , PV4.txInfoValidRange = timeRange
             , PV4.txInfoRedeemers = plutusRedeemers
-            , PV4.txInfoData = PV3.unsafeFromList $ Alonzo.transTxWitsDatums (ltiTx ^. witsTxL)
+            , PV4.txInfoData = PV4.unsafeFromList $ Alonzo.transTxWitsDatums (ltiTx ^. witsTxL)
             , PV4.txInfoId = Conway.transTxBodyId txBody
             , PV4.txInfoVotes = Conway.transVotingProcedures (txBody ^. votingProceduresTxBodyL)
             , PV4.txInfoProposalProcedures =
@@ -663,7 +670,11 @@ instance EraPlutusTxInfo 'PlutusV4 DijkstraEra where
             , PV4.txInfoCurrentTreasuryAmount =
                 strictMaybe Nothing (Just . transCoinToLovelace) $ txBody ^. currentTreasuryValueTxBodyL
             , PV4.txInfoTreasuryDonation = transCoinToLovelace $ txBody ^. treasuryDonationTxBodyL
-            , PV4.txInfoSubTxIx = Nothing -- TODO thread the subtx index here
+            , PV4.txInfoSubTxIx =
+                withBothTxLevels
+                  ltiLevelInfo
+                  (const Nothing)
+                  (\case SubTxInfo (TxIx i) -> Just $ toInteger i)
             , PV4.txInfoWithdrawals = transTxBodyWithdrawals txBody
             , PV4.txInfoDirectDeposits = transTxBodyDirectDeposits txBody
             , PV4.txInfoAccountBalanceIntervals =
@@ -673,7 +684,38 @@ instance EraPlutusTxInfo 'PlutusV4 DijkstraEra where
             }
       Right $ \_ -> Right txInfo
 
-  toPlutusArgs = toPlutusV4Args
+  toPlutusArgs proxy LedgerTxInfo {..} sh txInfo plutusPurpose redeemerData = do
+    scriptPurpose <- toPlutusScriptPurpose proxy ltiProtVer sh plutusPurpose
+    let
+      maybeSpendingData = getSpendingDatum ltiUTxO ltiTx $ hoistPlutusPurpose toAsItem plutusPurpose
+    topTxInfo <-
+      withBothTxLevels
+        ltiTx
+        ( \tx ->
+            Just
+              <$> transTopTxInfo
+                proxy
+                ltiProtVer
+                ltiUTxO
+                ltiSystemStart
+                ltiEpochInfo
+                txInfo
+                (hoistPlutusPurpose toAsPurpose plutusPurpose)
+                sh
+                tx
+        )
+        (const $ pure Nothing)
+    let
+      scriptInfo =
+        scriptPurposeToScriptInfo scriptPurpose (transDatum <$> maybeSpendingData) topTxInfo
+    pure $
+      PlutusV4Args $
+        PV4.ScriptContext
+          { PV4.scriptContextTxInfo = txInfo
+          , PV4.scriptContextRedeemer = Babbage.transRedeemer redeemerData
+          , PV4.scriptContextScriptInfo = scriptInfo
+          , PV4.scriptContextScriptHash = transScriptHash sh
+          }
 
   toPlutusTxInInfo _ = transTxInInfoV4
 
@@ -711,7 +753,7 @@ transTxOutV4 ::
   Either (ContextError era) PV4.TxOut
 transTxOutV4 txOutSource txOut = do
   let
-    val = transValue $ txOut ^. valueTxOutL
+    val = Alonzo.transValue $ txOut ^. valueTxOutL
     referenceScript = transReferenceScript $ txOut ^. referenceScriptTxOutL
     datum =
       case txOut ^. datumTxOutF of
@@ -742,8 +784,8 @@ transTxOutV4 txOutSource txOut = do
       }
 
 transTxBodyWithdrawals ::
-  DijkstraEraTxBody era => TxBody l era -> PV4.Map PV4.AccountId PV4.Lovelace
-transTxBodyWithdrawals txb = transMap transAccountAddressToAccountId transCoinToLovelace withdrawals
+  DijkstraEraTxBody era => TxBody l era -> PV4.Map PV4.Credential PV4.Lovelace
+transTxBodyWithdrawals txb = transMap transAccountAddress transCoinToLovelace withdrawals
   where
     Withdrawals withdrawals = txb ^. withdrawalsTxBodyL
 
@@ -785,14 +827,11 @@ transTxBodyRequiredTopLevelGuards ::
   DijkstraEraTxBody era => TxBody l era -> PV4.Map PV4.Credential (Maybe PV4.Datum)
 transTxBodyRequiredTopLevelGuards txb = transMap transCred (fmap transDatum . strictMaybeToMaybe) requiredGuards
   where
-    requiredGuards = txb ^. requiredTopLevelGuardsL
-
-transAccountAddressToAccountId :: AccountAddress -> PV4.AccountId
-transAccountAddressToAccountId (AccountAddress _ (AccountId c)) = PV4.AccountId $ transCred c
+    requiredGuards = txb ^. requiredTopLevelGuardsTxBodyL
 
 transTxBodyDirectDeposits ::
-  DijkstraEraTxBody era => TxBody l era -> PV4.Map PV4.AccountId PV4.Lovelace
-transTxBodyDirectDeposits txb = transMap transAccountAddressToAccountId transCoinToLovelace deposits
+  DijkstraEraTxBody era => TxBody l era -> PV4.Map PV4.Credential PV4.Lovelace
+transTxBodyDirectDeposits txb = transMap transAccountAddress transCoinToLovelace deposits
   where
     DirectDeposits deposits = txb ^. directDepositsTxBodyL
 
@@ -806,7 +845,7 @@ transAccountBalanceInterval = \case
 transAccountBalanceIntervals :: AccountBalanceIntervals era -> PV4.AccountBalanceIntervals
 transAccountBalanceIntervals (AccountBalanceIntervals balanceIntervals) =
   PV4.AccountBalanceIntervals $
-    transMap transAccountAddressToAccountId transAccountBalanceInterval balanceIntervals
+    transMap (transCredToAccountId . unAccountId . aaId) transAccountBalanceInterval balanceIntervals
 
 transTxBodyGuards :: DijkstraEraTxBody era => TxBody l era -> [PV4.Credential]
 transTxBodyGuards txb = fmap transCred . F.toList $ txb ^. guardsTxBodyL
@@ -825,31 +864,100 @@ scriptPurposeToScriptInfo sp datum topInfo = case sp of
   PV4.Proposing _ ix proc -> PV4.ProposingScript ix proc
   PV4.Guarding _ ix -> PV4.GuardingScript ix topInfo
 
-toPlutusV4Args ::
-  ( AlonzoEraUTxO era
-  , EraPlutusTxInfo PlutusV4 era
+transValidityInterval ::
+  Inject (Alonzo.AlonzoContextError era) (ContextError era) =>
+  Proxy era ->
+  EpochInfo (Either Text) ->
+  SystemStart ->
+  ValidityInterval ->
+  Either (ContextError era) PV4.POSIXTimeRange
+transValidityInterval era epochInfo systemStart (ValidityInterval from to) = do
+  let transSlot = transSlotToPOSIXTime era epochInfo systemStart
+  pFrom <- traverse transSlot from
+  pTo <- traverse transSlot to
+  pure $ PV4.POSIXTimeRange (strictMaybeToMaybe pFrom) (strictMaybeToMaybe pTo)
+
+transTopTxInfo ::
+  forall era proxy.
+  ( EraUTxO era
+  , DijkstraEraTxBody era
+  , Inject (DijkstraContextError era) (ContextError era)
+  , ConwayEraPlutusTxInfo PlutusV4 era
   ) =>
-  proxy 'PlutusV4 ->
-  LedgerTxInfo era ->
+  proxy PlutusV4 ->
+  ProtVer ->
+  UTxO era ->
+  SystemStart ->
+  EpochInfo (Either Text) ->
+  PlutusTxInfo PlutusV4 ->
+  PlutusPurpose AsPurpose era ->
   ScriptHash ->
-  PV4.TxInfo ->
-  PlutusPurpose AsIxItem era ->
-  Data era ->
-  Either (ContextError era) (PlutusArgs 'PlutusV4)
-toPlutusV4Args proxy LedgerTxInfo {..} sh txInfo plutusPurpose redeemerData = do
-  scriptPurpose <- toPlutusScriptPurpose proxy ltiProtVer sh plutusPurpose
+  Tx TopTx era ->
+  Either (ContextError era) PV4.TopTxInfo
+transTopTxInfo proxy pv utxo systemStart epochInfo _plutusTxInfo plutusPurpose sh tx = do
   let
-    maybeSpendingData = getSpendingDatum ltiUTxO ltiTx $ hoistPlutusPurpose toAsItem plutusPurpose
-    -- TODO TopTxInfo should be set if this is a top-level transaction
-    scriptInfo = scriptPurposeToScriptInfo scriptPurpose (transDatum <$> maybeSpendingData) Nothing
-  pure $
-    PlutusV4Args $
-      PV4.ScriptContext
-        { PV4.scriptContextTxInfo = txInfo
-        , PV4.scriptContextRedeemer = Babbage.transRedeemer redeemerData
-        , PV4.scriptContextScriptInfo = scriptInfo
-        , PV4.scriptContextScriptHash = transScriptHash sh
+    txb = tx ^. bodyTxL
+    mkLedgerTxInfo :: Tx SubTx era -> TxIx -> LedgerTxInfo era
+    mkLedgerTxInfo subTx txIx =
+      LedgerTxInfo
+        { ltiUTxO = utxo
+        , ltiTx = subTx
+        , ltiSystemStart = systemStart
+        , ltiProtVer = pv
+        , ltiLevelInfo = SubTxInfo txIx
+        , ltiEpochInfo = epochInfo
         }
+  let
+    subTxs = txb ^. subTransactionsTxBodyL
+  subTxInfosPartial <-
+    zipWithM
+      ( \stx txIx ->
+          first (inject . SubTxContextError (txIdTx stx))
+            . unPlutusTxInfoResult
+            . toPlutusTxInfo proxy
+            $ mkLedgerTxInfo stx txIx
+      )
+      (toList subTxs)
+      [TxIx i | i <- [0 ..]]
+  subTxInfos <- traverse ($ plutusPurpose) subTxInfosPartial
+  let
+    datums =
+      foldMap
+        ( \subTx ->
+            case Map.lookup (ScriptHashObj sh) $ subTx ^. bodyTxL . requiredTopLevelGuardsTxBodyL of
+              Just (SJust v) -> Map.singleton (txIdTx subTx) v
+              _ -> mempty
+        )
+        subTxs
+  pure $
+    PV4.TopTxInfo
+      { topTxInfoSubTransactions = subTxInfos
+      , topTxInfoStartingAccountBalanceIntervals =
+          transAccountBalanceIntervals $ txb ^. accountBalanceIntervalsTxBodyL
+      , topTxInfoDatums = transMap transTxId transDatum datums
+      , topTxInfoSimplified =
+          -- TODO implement TopTxInfoSimplified translation
+          PV4.TopTxInfoSimplified
+            { ttisWithdrawals = AssocMap.empty
+            , ttisVotes = AssocMap.empty
+            , ttisValidRange = PV4.POSIXTimeRange Nothing Nothing
+            , ttisTxCerts = mempty
+            , ttisTreasuryDonations = PV4.Lovelace 0
+            , ttisScriptPurposes = mempty
+            , ttisReferenceInputs = mempty
+            , ttisProposalProcedures = mempty
+            , ttisOutputs = mempty
+            , ttisMints = PV4.emptyMintValue
+            , ttisInputs = mempty
+            , ttisIds = mempty
+            , ttisGuards = mempty
+            , ttisDirectDeposits = AssocMap.empty
+            , ttisData = AssocMap.empty
+            , ttisCurrentTreasuryAmount = Nothing
+            , ttisBurns = PV4.emptyMintValue
+            , ttisRequiredTopLevelGuards = mempty
+            }
+      }
 
 transTxId :: TxId -> PV4.TxId
 transTxId (TxId h) = PV4.TxId $ transSafeHash h
@@ -866,7 +974,7 @@ transPlutusPurposeV4 ::
 transPlutusPurposeV4 proxy pv (transScriptHash -> sh) = \case
   DijkstraSpending (AsIxItem _ (TxIn txId (TxIx ix))) ->
     pure . PV4.Spending sh $ PV4.TxOutRef (transTxId txId) (toInteger ix)
-  DijkstraMinting (AsIxItem _ pId) -> pure . PV4.Minting sh $ transPolicyID pId
+  DijkstraMinting (AsIxItem _ pId) -> pure . PV4.Minting sh $ Alonzo.transPolicyID pId
   DijkstraCertifying (AsIxItem ix cert) ->
     pure $ PV4.Certifying sh (toInteger ix) (transTxCertV4 proxy pv cert)
   DijkstraWithdrawing (AsIxItem _ (AccountAddress _ (AccountId c))) ->